Allow for toggling of layers
This adds functionality that removes untoggled toggle-able layers from the builtLayers property of the map-by-id component. Toggles are config-driven with this schema:
```
map:
  toggles:
  - layerId: id-of-layer
    type: county
  - layerId: id-of-layer-2
    type: region
```

Toggle types are generated basesd simply on the name of the type in the toggles - they’re just capitalized in the template, so there can be any type of switch. It simply swaps out different layers.