Change property names; Add back LayerTitle;
Instead of replacing the layers property of our map narrative configurations, this commit changes that to add a new property called `mapboxLayers`, which live on their own as specially formatted objects that are compatible with mapbox gl style spec. This way, we can always expect to get what we see in the YAML configs, since nothing is being done to those structures.